Update build.gradle
Our internal nexus maven repository was previously listed *before* the github.io maven repo, which causes android studio to hang when loading the project, as the dependencies cannot be met from the nexus repository without the vpn. 

Until this is accepted, the learning app as cloned is broken for users.